+++ src/emu/winapi/helper.rs
@@ -2,13 +2,13 @@ use lazy_static::lazy_static;
 use std::sync::Mutex;
 
 lazy_static! {
-    static ref HANDLERS:Mutex<Vec<u32>> = Mutex::new(vec![0;0]);
-    static ref SOCKETS:Mutex<Vec<u32>> = Mutex::new(vec![0;0]);
+    static ref HANDLERS:Mutex<Vec<u64>> = Mutex::new(vec![0;0]);
+    static ref SOCKETS:Mutex<Vec<u64>> = Mutex::new(vec![0;0]);
 }
 
 
-pub fn handler_create() -> u32 {
-    let new_handle:u32;
+pub fn handler_create() -> u64 {
+    let new_handle:u64;
     let mut handles = HANDLERS.lock().unwrap();
 
     if handles.len() == 0 {
@@ -22,7 +22,7 @@ pub fn handler_create() -> u32 {
     return new_handle;
 }
 
-pub fn handler_close(hndl:u32) -> bool {
+pub fn handler_close(hndl:u64) -> bool {
     let mut handles = HANDLERS.lock().unwrap();
     let idx = match handles.iter().position(|h| *h == hndl) {
         Some(i) => i,
@@ -39,7 +39,7 @@ pub fn handler_print() {
     }
 }
 
-pub fn handler_exist(hndl:u32) -> bool {
+pub fn handler_exist(hndl:u64) -> bool {
     let handles = HANDLERS.lock().unwrap();
     match handles.iter().position(|h| *h == hndl) {
         Some(_) => return true,
@@ -48,8 +48,8 @@ pub fn handler_exist(hndl:u32) -> bool {
 }
 
 
-pub fn socket_create() -> u32 {
-    let new_socket:u32;
+pub fn socket_create() -> u64 {
+    let new_socket:u64;
     let mut sockets = SOCKETS.lock().unwrap();
 
     if sockets.len() == 0 {
@@ -66,7 +66,7 @@ pub fn socket_create() -> u32 {
     return new_socket;
 }
 
-pub fn socket_close(sock:u32) -> bool {
+pub fn socket_close(sock:u64) -> bool {
     let mut sockets = SOCKETS.lock().unwrap();
     let idx = match sockets.iter().position(|s| *s == sock) {
         Some(i) => i,
@@ -76,7 +76,7 @@ pub fn socket_close(sock:u32) -> bool {
     return true;
 }
 
-pub fn socket_exist(sock:u32) -> bool {
+pub fn socket_exist(sock:u64) -> bool {
     let sockets = SOCKETS.lock().unwrap();
     match sockets.iter().position(|s| *s == sock) {
         Some(_) => return true,
